rename: Printer, UserInputReceiver를 io 패키지로 이동
