+++ src/main/java/lotto/io/Printer.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.io;
 
 import lotto.game.WinningLottoCounter;
